Changed robot.py to work with the actual robot
